Move worker code to util function
